Added contact
I've added the contact and footer, the only problem I am having is correctly setting up the form, and also connecting it so that it sends the message via email to one of my emails.

Once I learn the skills required in order to do this, I will implement it into this web application.